Start generating FF-parser trivia nodes from schema
Summary:
We generate token and ast code from the schema; I am considering adding new trivia kinds and so it would be nice to have the trivia code generated as well.

This is the first step, just building the infrastructure and generating the schema.

Note that many of the changes were the result of having to use a map that omits the trailing comma in the json representation of the schema. Languages like Hack which are generous in their support for trailing commas save on much irritating scutwork.

Reviewed By: jamesjwu

Differential Revision: D4354938

fbshipit-source-id: 15bc44e4c9500ff13492c73a4a07a6cdf4bc5623