Use canonical names when adding files to rename in global naming
Summary:
Previously, we only used the error file's name when adding files to the failed naming list rather than the canonical name, meaning that we were finding incorrect positions and not adding enough files to the list of errored files when the files differed only in capitalization.

This bug seems to predate lazy decl, as described in t14949125.

Reviewed By: dabek

Differential Revision: D4373103

fbshipit-source-id: 64380c0bc56d52ae74220b1fcfc69f87bdeba88a