Refactor/add check to enforce that we do not try and invoke a class method on an object which is not an instance of the implementing class
Summary:
This code exists in upstream PHP to enforce this same restriction.

https://github.com/php/php-src/blob/e52d2b88e4871f2e82fc26986d97417e76af8127/ext/reflection/php_reflection.c#L3283-L3290
https://github.com/php/php-src/blob/e52d2b88e4871f2e82fc26986d97417e76af8127/ext/reflection/php_reflection.c#L3193-L3196

A similar check already existed in invokeArgs but it enforced that you could only call the method if the object passed in was a subclass of the class being inspected, rather than the implementing class, which is incorrect. It is valid to have a class X with a child class Y and use ReflectionClass on Y with an object of type X to invoke a method declared on X (we do this in `hphp/test/slow/late_static_binding/1887.php`).

The main challenge here was that in order to do this validation you need two pieces of information:

1. The object
2. The declaring classname for the method (ie: the name used when instantiating the ReflectionClass or a subclass depending on the hierarchy)

In order to actually execute the code you need two pieces of information:

1. The object
2. The provided class name (ie: the one used when instantiating the ReflectionClass)

Reviewed By: markw65

Differential Revision: D4372646

fbshipit-source-id: 25f842e1fe89777e51044970017aca1724fc078b