Lex whitespace / newline trivia within XHP content blocks
Summary:
drphil points out to me that XHP bodies are not merely unstructured blobs of text, but rather that whitespace is special.

Specifically, any number of newlines, tabs, spaces, etc, in a row are semantically the same as a single space.  The text and tags inside XHP can be spaced arbitrarily. But in the lexer we were treating spaces as just part of the text.

Obviously this is of relevance to pretty printers, code analyzers, and so on.

We now treat spaces, etc, within an XHP body as trivia, and non-whitespace text as XHP body tokens.

Reviewed By: KendallHopkins

Differential Revision: D4378281

fbshipit-source-id: d5bea2947fff82454d61666859de3d414a96602f