Fix a buffer overrun in pcre_literal_data
Summary:
According to the man page, strchr(s, '\0') returns a pointer to s's
nul terminator (a surprise to me), so the loop doesn't terminate until
it finds a bogus character. In one of the cores I'm looking at, its
clear that the loop walked off the end of the string and through to
the start of the next 2M slab, where it segfaulted.

I think this means the current version only finds strings that end
with '$'.

Reviewed By: mxw

Differential Revision: D4389821

fbshipit-source-id: d5cf70b19047884d8f68a0844a890aa606bc5f1d