test: cycle_debug_backtrace: Fix hard coded value.
The expected test output for test
hphp/test/slow/async/cycle_debug_backtrace.php
contains a hard coded string length for the expected
stack. As the stack contains object IDs, which happen
to be virtual addresses in HHVM, expecting a specific
length of a string output is not correct.
In fact this test never matched on arm64.

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>