improve range() behavior to match php7
Summary:
Previously, HHVM was mimicking PHP5 implementation of range(): even when range boundaries are integers, they are internally converted to floating point. Due to precision loss, this can cause range() to return incorrect results or even to get into infinite loop (if low + step == low due to precision loss).

This is fixed in PHP7, so let's fix it in HHVM too:
- When range boundaries are integers, use integers for iteration. This avoids precision loss altogether
- When range boundaries are floating point, use integer counter for iteration, which prevents getting into infinite loop due to precision loss

Reviewed By: markw65

Differential Revision: D4390506

fbshipit-source-id: cb45c6d37ac26e10647cd718e08ca090020f06e5