IROPs for packed ArrayData should accept both packed array and vec.
Summary:
We had LdPackedArrayElemAddr and LdVecElemAddr, which differed only in the
type of the operand; merge them together. Make CheckPackedArrayBounds also
support vecs. To support these changes, make the AK() syntax in the
ir.specification usable within the S() syntax; this lets us declare ops that
take either a packed array or a vec.

Reviewed By: ricklavoie

Differential Revision: D4395817

fbshipit-source-id: 6b7d3ef021be65f65a1579dcc0811fc9a0de2ebc