Improve cast heuristic
Summary:
Note that this is a change to the original parser, not the FF parser.

The Hack parser must disambiguate between casts and parenthesized expressions:

    (10)-2   // definitely parenthesized expression
    (int)-2  // definitely cast
    (blah)something // Which?

To disambiguate the last case, it checks the following token; if it is something like [, ], ;, etc, then it cannot be a cast.

Unfortunately ) was left off the list, so the expression (false && (true)) gives an "expected expression" error because (true) is assumed to be a cast.

I've added ) to the list.

The full-fidelity parser already has a somewhat more advanced heuristic based upon the heuristic used by C#; see the comments in the FFP sources for details.

Differential Revision: D4404439

fbshipit-source-id: 3102f31192e6485923ceca30f7f03f7a7bd96b33