Preserve fast call abi rules for dec ref.
Summary:
Previously the dec ref stub made use of cmplim and declm, which when
lowered on AArch64 would use virtuals. This problem was originally
addressed with special vasm variations which when lowered/emitted for
AArch64 would avoid the use of new virtuals.

The new solution introduces a flag, `fast`, in the `Vunit` which indicates to
any relevant lowerers that the platform's rAsm should be used instead of
new virtuals.
Closes https://github.com/facebook/hhvm/pull/7393

Differential Revision: D3938038

Pulled By: mxw

fbshipit-source-id: dc79b39197911ab1a84f60ca887db687728cf47e