Propagate detailed exception message from ExecutionContext::evalPHPDebugger to the XDebug debugger.
Summary:
If an exception is encountered while evaluating a PHP expression from a remote debugger, the exception details (which contain useful information about what actually went wrong and the stack trace of the current VM frame) are swallowed and not propagated out to xdebug. Instead a generic error code and message like "error evaluating code" is displayed.

This change captures the detailed error message into the XDebugExn thrown by the caller of evalPHPDebugger, and gives the full text to be displayed in Nuclide.

Reviewed By: yinghuitan, paulbiss

Differential Revision: D4396889

fbshipit-source-id: cfd587879b5f5bb9f24c5fe1267199137e4dc20d