Add tracking of equivalencies between locals and stack values to HHBBC
Summary:
Add the ability for HHBBC to track (in a somewhat limited manner) value
equivalencies between locals and stack elements. Use these equivalencies to
optimize the MemoGet and MemoSet ops. The idea is if GetMemoKeyL is an identity
operation on function parameters, then there's no need to use a separate set of
locals for the memoization keys. Instead we'll infer that the locals
representing the function parameters and the redundant set of locals are
identical, and rewrite MemoGet/MemoSet to use the parameters directly. Then the
redundant set of locals will be removed since they're dead. The equivalency
tracking is primitive at this point (it only deals with CGetL and SetL), since
that's all that's needed for the above optimization.

Reviewed By: swtaarrs

Differential Revision: D4376975

fbshipit-source-id: 265cf2137b5a4339f03c678e40fddf4970864d55