Fixing bug in hphpd_auth_token() extension
Summary:
This diff fixes an issue introduced in D3725724 which causes a race condition between the polling server and the auth call. The fix is to add a lock at the beginning of `DebuggerProxy::requestAuthToken()`.

paulbiss graciously helped me debug this -- would never have figured it out by myself! :)

Note that hphpd_auth_token() is currently not called by any production code, so there is low risk with modifying this function.

Reviewed By: paulbiss

Differential Revision: D4416314

fbshipit-source-id: 113fab8fa7df23c4d29d76209c6a8f1a8b3b0d76