Fix bug in HHVM xdebug_sprintf error path
Summary:
This change fixes a bug I found in xdebug_sprintf when the underlying call to vsnprintf() fails.

Per man page, vsnprintf returns a negative integer on an error (but not when the buffer is too small). The current code would continually realloc the buffer, doubling its size each time in a loop until the process runs out of memory and dies in this case, or loop forever if the malloc/realloc implementation doesn't throw on out of memory.  If hit, this could cause HHVM to either crash or hang (depending on malloc's beahvior).

Instead, I'm freeing the allocated buffer (if any) and returning nullptr here on error.

Reviewed By: yinghuitan, johnislarry

Differential Revision: D4411966

fbshipit-source-id: d055a0c5e984d144992ef8ba6610607a7ad1bd18