Fix base type update for F<minstr> instructions
Summary:
F<minstr> just runs both versions of the minstr, and merges the state;
but it doesn't merge the types of the base.

This means that eg FDim can convert a Cell to an InitCell, or even a
Null to an array or stdClass; and in some cases, that type can be
written back to the base local.

In the new test case, on the first iteration, we have an FDim, and the
type of $row is updated to InitCell, which then gets picked up as a
use-var by the lambda. On a later iteration, insert has been resolved
to a method on Table, the FDim has become Dim; FPassC, and the type is
left at Cell, resulting in an assertion because we try to widen a type
in the index.

Reviewed By: ricklavoie

Differential Revision: D4420292

fbshipit-source-id: 1a78834292e5912b9b87147f2aa33cf60fb7bfa0