More fixes for mergePaths
Summary:
We could setup an arrayChain on the define side, but not the
other. Since we're mreging the two branches, we need to resolve it. In
addition, the asserts I added with D4420292 were too strong; the
local-related fields are only meaningful when the type for
BaseLoc::Frame and BaseLoc::LocalArrChain.

Reviewed By: ricklavoie

Differential Revision: D4422084

fbshipit-source-id: 0880a9d577fdf238ffae7345f107957ee9426313