Improve lazy recheck batching behaviour
Summary:
The algorithm aims to do a lazy recheck "when we are idle", which is right now defined as "100 ms have passed since last edit command". But in practice, when you are typing, the editor doesn't send a stream of EDIT_FILE, EDIT_FILE, EDIT_FILE, EDIT_FILE... it's at least EDIT_FILE, AUTOCOMPLETE, EDIT_FILE, AUTOCOMPLETE, EDIT_FILE ... with some more commands possibly thrown in. If you are working on a big (few thousand lines) file, the autocomplete itself can exceed 100ms, and you end up triggering a recheck on every keystroke which is what we wanted to avoid in the first place.

Changing "is_idle" definition to include "does not have any commands pending".

Reviewed By: alexchow

Differential Revision: D4417074

fbshipit-source-id: c629c4eb372eecd3fc9453e0c8b878547c7f50d3