Implement fallthrough directive
Summary:
We will eventually replace the comment-based // FALLTHROUGH directive with a proper keyword that can appear at the end of a switch section.

This diff gets the parsing of this pseudo-statement right in the FF parser; when eventually it is the "real" parser we can do code mods to eliminate the comment-based syntax and use the directive.

Note that the error message -- like most of the FFP error messages -- is reported in the wrong location (the newline!).  I'll be working on improving this soon.

Reviewed By: oulgen

Differential Revision: D4414807

fbshipit-source-id: 911239f1e8d124d48ed928dd7bcb8d3719e951eb