test: cycle_debug_backtrace: Fix hard coded value.
Summary:
The expected test output for test
hphp/test/slow/async/cycle_debug_backtrace.php
contains a hard coded string length for the expected
stack. As the stack contains object IDs, which happen
to be virtual addresses in HHVM, expecting a specific
length of a string output is not correct.
In fact this test never matched on arm64.

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>
Closes https://github.com/facebook/hhvm/pull/7597

Reviewed By: aorenste, paulbiss

Differential Revision: D4429804

Pulled By: mofarrell

fbshipit-source-id: e5ce85149b3a6f2a2d6f53250ebb98045856d3d7