Wrap variadic parameter declarations in descriptive type
Summary:
The FF parser would take a parameter list like

    (int $x, ...)

and create a parse tree where the first list item is a parameter declaration (obviously) and the second list item is an ellipsis token.

Kendall reasonably pointed out that having a list where the two things that can be in it are a highly structured node type -- parameter declaration -- and a single ellipsis token -- is a poor user experience in the API.  I agree.

We now wrap ellipses in function, lambda and function type parameter lists in a "VariadicParameter" structure that has one child, an ellipsis token.

Reviewed By: KendallHopkins

Differential Revision: D4432344

fbshipit-source-id: 35270b3392caf06d6b2b6b64cbdfa9c7a782b486