test: Adding missing valid output variant to phpunit_mock.php.
phpunit_mock.php is expected to provide a string of this format:
Time: %d second%A, Memory: %s

However, it is possible that the test outputs this (observed on a debug
build on arm64 Ubuntu Xenial machine, which was under heavy load by
other tasks):
Time: 01:13, Memory: 22.92Mb

This leads to a failed test.
The reason is, that phpunit.phar:secondsToTimeString() creates several
time string variants (depending on the duration).
The expected test output assumes that the runtime is always below
one minute. As observed this assumption might not be always true.

This patch allows the test to take up to one hour, which should be
a reasonable timeframe.

Verification of this change has been achieved starting stress(1)
in background:
stress --cpu $(nproc) --io $(nproc) --vm $(nproc) --hdd $(nproc) --timeout 120s

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>