Fixing pathing bug with hphpd_auth_token() script
Summary:
This diff only affects the behavior of the debugger when invoking the function `hphpd_auth_token()`. All other functionality remains the same.

A bug in D3725724 caused the token script to be triggered by this command to only be found if the debugger was running in `~/www`. This fix allows for the debugger to be run anywhere.

When the debugger connects to a sandbox, the default sandbox path is sent from the server to the client, which the client then uses the construct the full token script path.

There could potentially be a discrepancy when the debugger connects to a sandbox which does not set the path correctly for when the client triggers `hphpd_auth_token()`, but this will simply cause the command to return the empty string.

When attempting to land, it looks like there are test failures that are not my fault. See:
- https://fb.facebook.com/groups/734933446555273/permalink/1293064390742173/
- https://www.facebook.com/groups/1887034218249824/permalink/1918538345099411/

So I'm going to override the unit test failures to get the land through with:

Reviewed By: paulbiss

Differential Revision: D4431538

fbshipit-source-id: b8b1d625f1829f4f9cf356b7af4124e1fc0a8aca