Improve parse tree representation of encapsulated strings.
Summary:
We represent the interior of an encapsulated string as a list of text sections.  Previously if we had something like

    "abc{$x}def"

that would be represented as a list:

    text head   "abc{
    expression  $x
    text tail   }def"

But the { and } are not part of the text of the string, so this is somewhat confusing.  We now represent this as

    text head          "abc
    braced expression  {,  $x,  }
    text tail          def"

We make similar changes to the deprecated form:

    "abc${x}def"

is now

     text head         "abc
     dollar            $
     braced expression {,  x,  }
     text tail         def"

I've also added comments clarifying how trivia is assigned to fragments.  For example, in

    "abc ${ x } def"

there are four spaces. The first is part of the text fragment; it is not trivia at all. The second is trailing trivia of the {. The third is trailing trivia of x, and the fourth is part of the tail text fragment.

Not in this change: we might consider similarly calling out the leading and trailing double quotes as special string regions. Similarly for detected escape sequences.

Also not in this change:  The expression inside "{$x}" is parsed as a variable, but in "${x}" it is parsed as a name; the original parser parses it as a variable.  I'll need to fix that up at some point.

Differential Revision: D4430418

fbshipit-source-id: caf5a5eb68a1f58c6a18e3d3914beac534bfd0da