Remove incorrect eval stack frame logic
Summary: I added this logic in a long time ago to distinguish "eval" frames from "file" frames, but I've since come to realize that it's not right, because pseudomain frames don't always indicate that we've started an eval, for instance any `include` will cause a pseudomain frame to be on the stack.  This diff removes the logic.

Reviewed By: paulbiss

Differential Revision: D4446088

fbshipit-source-id: c2243fbac3d8826191662da0ef554f3700ec8b88