Fix crash in original Hack parser involving extra ellipsis
Summary:
The program fragment

    function x(... int ... $x) {}

crashes the parser; the code is *asserting* that it is impossible for there to be ellipses in both places, but in fact this is a rule of the language that should be detected to produce an error, not a logical invariant.

I've modified the parser so that the invariant is maintained and an error is produced for this program fragment. (The error is "expected variable").

Differential Revision: D4443342

fbshipit-source-id: 6bb6384928b81a91837d5c55d61372a0cb11868b