Disallow $$ in locations where variable is expected
Summary:
While investigating how Hack parses parameter lists I discovered an oddity. The Hack parser allows this program:

     class C {
       public int $$;
       function foo($$) {
         try {
           $x = $y |> function blah() use($$) { };
         }
         catch (Bar $$) { }
       }
     }

All of those uses of $$ are, with the possible exception of the "use" clause, nonsense. All are accepted by the Hack parser. All are rejected by the HHVM parser.

For reasons unknown to me there is a helper method "variable" in the Hack parser which is *only* called in the four situations above: class variables, formal parameters, use clauses and catch clauses, and which *specifically* accepts $$.

I've fixed the helper method so that it only accepts local variables, as it should.

Reviewed By: andrewjkennedy

Differential Revision: D4445087

fbshipit-source-id: 81fe4bfccfa560858f26d49f8dde43dfcbdd5e33