Properly type invariant_violation to be sprintf-style
Summary:
The lack of this is bugging me. I've spent a bunch of time trying to clean this up, but I can only work on it intermittently and it requires setting up a custom hh_client/hh_server, which is annoying and slow (I need to wait for hh_server to warm up) so there isn't a ton of progress.

The plan is to mute all of the existing badness with HH_FIXME, then come back later and clean it up (and get other people to work on it too) I'm in the middle of using fixme.par and a version of hh_client to write diffs to add those. I won't land this until the weekend, after I've landed those diffs and confirmed trunk is clean.

Differential Revision: D4414278

fbshipit-source-id: 00e9199508add966b6cec618cd290633e52f0fba