"Forced conservative" types should include base classes
Summary:
The type-scanner generator has machinery to always conservative scan
instantiations of certain template types. This is used for types like
std::optional, which cannot be automatically scanned, and which we don't want to
extend to add custom annotations.

We want to do the same for std::function, but it doesn't work. The reason is
because the forced conservative machinery does not extend the forcing to the
type's base classes. Fix this. Now if you mark a type as forced conservative,
that will extend to that type's base classes as well.

Now that this works, add std::function to the forced conservative list.

Reviewed By: edwinsmith

Differential Revision: D4451665

fbshipit-source-id: f074f7a23f647b676b54ac57f7c6722cc20730f9