Use a static cast correctly in asm-ppc64
Summary:
In MSVC it's an error to use `reinterpret_cast` to cast an integer value to the type it already is.
This switches the `reinterpret_cast` to a `static_cast`.
Technically the cast isn't needed at all, but it makes it explicit what the type being written is, so meh.

Reviewed By: paulbiss

Differential Revision: D4444865

fbshipit-source-id: def0d722e48960761bc1595e633d55e9b7ca9fce