test: make member reflection optional.
Member reflection is an optional feature, but when not used
makes three test cases fail:

* config_overrides.php
* tier_overrides_default.php
* tier_overrides_tier.php

The reason of the fail is, that init_member_reflection() creates
an additional output line, which warns about missing embedded data.

This patch makes the tests aware of that (optional) line
and thus fixes all three tests.

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>
Signed-off-by: Christoph Müllner <christophm30@gmail.com>