Real fix for wasPEI assert in mergePaths
Summary:
I removed an assert that was failing when we had a hack array with a
known key; but I think the assert was correct. Given an FPass*, the
define side, "b()" in mergePaths, can always throw, but we weren't
merging env.flags, so once a() cleared the wasPEI flag, it was never
reset.

Reviewed By: ricklavoie

Differential Revision: D4459688

fbshipit-source-id: a4de894b0b568fcaa7064914569cb8058c83c1ee