Fix Hack parser to disallow trailing comma in variadic parameter
Summary:
While investigating why the FF parser parses variadic parameters incorrectly, I discovered that the original parser parses them incorrectly in a great many ways.

I am making a series of small changes to bring the original parser in line with the HHVM parser. Once that is done, then I'll get the FF parser correct too, and update the spec.

The Hack parser incorrectly allows a trailing comma:

    function foo ( ... , )

This should be an error.

It is now an error, but unfortunately, due to other problems in the parameter parsing, the error given is not "expected right paren" or "expected variable" -- the two things that could come here legally -- but rather "expected type", which is not legal.  The parser thinks that

    function foo ( ... string $x )

is legal, but it is not. I'll fix that in the next diff.

Differential Revision: D4463394

fbshipit-source-id: 86584d5f2b51e45a9ab7435ad4577db7b8920a81