Fix SetNewElem on keyset references
Summary:
Let's not crash when M base is a reference to a keyset and unpack the reference
like what array and vector helpers do.

Other helpers probably need to be audited as well. Perhaps a more general fix
that moves the responsibility of unpacking references to jit should be
considered.

Reviewed By: ricklavoie

Differential Revision: D4464372

fbshipit-source-id: 34b70f8b6f3509d6e940ebe4eb15de80e4bcd121