Continue fixing Hack parser treatment of variadics
Summary:
While investigating why the FF parser parses variadic parameters incorrectly, I discovered that the original parser parses them incorrectly in a great many ways.

I am making a series of small changes to bring the original parser in line with the HHVM parser. Once that is done, then I'll get the FF parser correct too, and update the spec.

In this diff: the Hack parser incorrectly allows a type hint after ... but before the variable:

    function foo ( ... string $x )

This should be an error. The correct syntax is

    function foo ( string ... $x )

This diff also fixes the problem mentioned in the previous diff: the error given for

    function foo ( ... , )

is now "expected variable" instead of "expected type".

There are still several problems with parsing ... that I will fix up in the next diff. In particular, note that function "param" is now always called with flag "variadic" equal to false.  I'll be eliminating that flag, and also disallowing

    function foo ( string $x ... )

Differential Revision: D4465366

fbshipit-source-id: 500b8887573571ecfede82ce40998960079bca5d