Continue to fix variadic parameters in Hack parser
Summary:
While investigating why the FF parser parses variadic parameters incorrectly, I discovered that the original parser parses them incorrectly in a great many ways.

I am making a series of small changes to bring the original parser in line with the HHVM parser. Once that is done, then I'll get the FF parser correct too, and update the spec.

In this diff: the Hack parser incorrectly allows a ... after a parameter declaration.

    function foo ( string $x ... )

This should be an error. The correct syntax is

    function foo ( string ... $x )

We still incorrectly allow

    function foo ( string & ... $x )

I'll fix that up in the next diff.

Differential Revision: D4466485

fbshipit-source-id: fa78c51f3742ceb181cdb2ef50375010b4456380