Disallow ref variadic parameters in Hack parser
Summary:
While investigating why the FF parser parses variadic parameters incorrectly, I discovered that the original parser parses them incorrectly in a great many ways.

I am making a series of small changes to bring the original parser in line with the HHVM parser. Once that is done, then I'll get the FF parser correct too, and update the spec.

In this diff: the Hack parser incorrectly allows

    function foo ( string & ... $x )

This is rejected by the HHVM parser.

This now has a new error message "A variadic parameter may not be passed by reference."  A couple thoughts on this error message.

First, any time you want to say that two things are incompatible, there's a tendency to assume that one of them is the intended state. Why is it not "A by-reference parameter must not be variadic?"  Or "A parameter cannot be both variadic and by reference", which assumes neither is intended.

In this case it seems most natural to assume that the writer intended "..." to mean that the parameter is variadic, so I assume that is the intended state.

Second, it is somewhat unfortunate that the error message is only produced for "&...$x" and not for, say "...&$x", which seems equally plausible. This produces the more prosaic "expected variable".  In an ideal world we would recognize lots of different variations on mixing and matching types, ellipses, variables and refs, and give explanatory messages for each. I'll explore this more in the full fidelity parser.

Differential Revision: D4467235

fbshipit-source-id: 2e7ba7a05dfd53d01e2816bfa45f65034034907f