Move the Environment portability header to the stdlib header
Summary: This completely kills the environment portability header, which was among the earliest, and was created before the general convention of splitting based on the header they are in, in the standard library, existed. The functions that were defined in it are expected to be in stdlib, so move them there.

Reviewed By: mzlee

Differential Revision: D4462502

fbshipit-source-id: c4549b9d1ea623644223db4365b81507f443ad5e