Typed AST: statements, top-level functions and constants
Summary:
Extend previous work on embedding types into expressions to apply this recursively inside statements. Statements themselves are not yet annotated - just the expressions embedded therein.

Top-level functions and constants are processed: see the signatures of `Typing.fun_def` and `Typing.gconst_def`.

Treatment of loops is naive. Because we iterate to get a fixed point type, we will generate essentially the same Typed AST each time around the loop, unnecessarily. Better would be to separate the computation of the type with the production of the typed AST.

Reviewed By: hubyrod

Differential Revision: D4449768

fbshipit-source-id: e23704cf9f2fdef6172a8108958c81d033dbfdff