Fix error messsages, comments and tests for variadic parameters and arguments
Summary:
Surprisingly, variadic parameters and arguments were being parsed correctly in the FF parser; it was in the original parser that they were parsed incorrectly.  In this diff I add comments indicating where the specification needs to be updated, some TODOs regarding the perhaps dodgy design choice to make a variadic parameter a kind of expression, and make the error reporting more consistently reported and worded.

I'll submit issues to the published specification for Rex to fix at a later date.

Differential Revision: D4471626

fbshipit-source-id: f56f3a0cf8b21ea88988c5393d385b63bd2a60d3