Watchman-based revision tracking in Hack Monitor Informant
Summary:
Watchman-based revision tracking in Hack Monitor Informant

See docs in code for more details.

In short,
Watchman tells us about state change events; we need query hg to
get corresponding SVN revisions for the HG revision in that event;
we keep queries in a map of deferred computations (Futures) so
all calls to the Informant are non-blocking; this map also doubles as
a cache (with no eviction policy :p) to trigger server restarts earlier.

Reviewed By: dlreeves, jamesjwu

Differential Revision: D4476118

fbshipit-source-id: 2717e74feae3334ca87d2c77f09928ad5c848a20