Fix errors in lexing hash comments
Summary:
Hack supports using a hash sign or two slashes to introduce a single-line comment. The lexer was getting it wrong in two ways.

First, it was skipping two characters before starting to look for the end of line, even for hash comments, meaning that a line with a single hash followed by newline was treating the line which followed as part of the comment.

Second, it was looking for "FALLTHROUGH", etc, inside hash comments, which is not desired.

Differential Revision: D4475560

fbshipit-source-id: 6d505b2bba0fe2a3209e2ad531472ca1d9f1cdaf