Add an interp,jit mode to test/run
Summary:
For the repo mode tests, we spend 90% of the time building the repo;
this allows us to build the repo once, and run the tests in both jit
and interp modes.

I was hoping to make each test run both ways and report separate
results, but that seemed to require more plumbing than I was prepared
to do. So this reports a single result for both modes (and if the
interp run fails, it doesn't try the jit run). This will make it
slightly harder to determine whats wrong - but I don't think its a big
deal.

Reviewed By: ricklavoie

Differential Revision: D4480639

fbshipit-source-id: 88a575cc8399d97f9123475880f638e90305adcd