v0.3.9 - Add lazy rules to allow for more flexible rule dependencies
Summary:
Allows for lazy rules. Perviously rules could only be dependent on rules that their contents contained

Example: the rule for deciding whether to split out f1s arguments can only depend on rules inside f1's arguments  (in this case it depends on the rules that govern f2's arguments:

  f1($a1, $a2, f2($b1, $b2)

Now we can set up other dependencies, so in the following case the rule that governs the split before ->funA1() is dependent on whether or not the arguments of getA() are split.

  $a::getA(
     $b,
     $c,
   )
     ->funA1()
     ->funA2()

Differential Revision: D4453846

fbshipit-source-id: b1bd2d6432db334a8a20a35a0e3da3a94c394d5c