Fix an assert in PrecomputedResults::get()
Summary:
The way the PrecomputedResults cache works is a little weird in that it
considers the entries involving invalid keys as not contained in the cache
(cf. ##contains()##), even after it's inserted back in the cache.  So we can't
assert that ##contains()## returns true after we insert the value until it's
marked as valid again.  This diff drops such assert, and replaces it with a
better one.

Differential Revision: D4463319

fbshipit-source-id: 0d490c310c2a01216801eb834854c5608e5a0af7