Fix precedence of ;: in FF parser
Summary: Due to a bug in the FF parser, expressions of the form a*self::b+c were parsing as a*((self::b)+c) instead of the obviously correct parse (a*(self::b))+c.  The solution was to stop trying to do too much when parsing the left side of :: and the :: itself, so that helper methods that understand precedence do the work.

Differential Revision: D4483132

fbshipit-source-id: 303764fdf9a165e4d9c050b2e7d3cf8ac12fd418