test: Make session IDs unique.
The session ID is used to create a filename of a new created file in /tmp.
If that is not unique, warning of the following form can be observed:

Warning: open(/tmp/sess_abtest, O_RDWR) failed: Permission denied (13)

This patch make the session ID unique, by calling uniqid() with the
defined string.

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>