Enable CRC32 instructions for hashing
Summary:
This patch enables CRC32 instructions for the Aarch64 architecture.
It adds new assembly routines to hash-crc-arm.S which utilize the
CRC32 instructions to accelerate string hashing.

It adds a new cmake flag, enabled with -DENABLE_AARCH64_CRC=1, in the
same fashion as the existing ENABLE_AVX flag. This is required since
there are not compile time flags which indicate whether or not the
given CPU truly supports the CRC32 instructions, as they optional.

It renames NO_SSECRC to NO_HWCRC to make the code processor agnostic.

It moves the x64 assembly from the hash-crc.S file to hash-crc-x64.S.

Lastly, it implements hash_int64 and hash_int64_pair using the CRC
instructions.
Closes https://github.com/facebook/hhvm/pull/7556

Differential Revision: D4357501

Pulled By: mxw

fbshipit-source-id: 5b5e363e0002742f0897033ed53468f5dfe1f19b