Move ext_watchman into open-source
Summary:
As ext_watchman has an unusual dependency on the Watchman client library it is marked as IMPLICIT in the cmake build. This means it will only block the build if it's forced to be enabled. If the client is available the extension will be built unless it is force disabled.

For some reason using `std::make_unique()` caused compiler errors with the cmake build so I've just switched to explicitly using `std::unique_ptr()` + `new`.

Some minor tweaks to the test output data as the cmake build doesn't seem to demangle C++ symbols in exception outputs (it's probably possible to enable but it's not the default).

Reviewed By: paulbiss

Differential Revision: D4319140

fbshipit-source-id: 390bea97ee3b2d98b55c7c414eb73ac0bcf3e25c