Make ext_watchman test only alert devx_www
Summary:
FACEBOOK

The ext_watchman test is a bit flakey but I'm having trouble repro'ing this. I suspect the problems are transient external failures relating to Watchman which I can't really control. After multiple hours of running in a loop on my devserver I got one failure which was of this nature.

This diff introduces a Faebook-only test which runs the real/open-source ext_watchman test but always returns "pass". All failures are instead reported directly to devx_www by creating/bumping a task. This way we can figure what's going on if there is a consistent problem without bothing the HHVM team. The original/open-source test can still run independently but we will disable it in Sandcastle internally.

Reviewed By: markw65

Differential Revision: D4447077

fbshipit-source-id: 85ca223823340317f543abc9d83c5000f949d54d