Skip inlined frames when attributing the cost at enter hooks
Summary:
Recent inliner improvements revealed discrepancy in Xenon cost attribution.
If a function A inlined a call to function B, which called non-inlined function
C, then the cost incurred by function A prior to the B call will be observed
during C's enter hook (we don't execute event hooks for inlined functions).

The current implementation will thus incorrectly attribute the cost to the B,
which is a parent of C. This diff skips inlined frames so that the cost will
get properly attributed to A.

Reviewed By: paulbiss

Differential Revision: D4492159

fbshipit-source-id: 7b85494201771d6d5095bcdcce71f4350b1b43e8