Fix associativity of ?: operator in FF parser
Summary:
The FF parser was incorrectly parsing nested ?: operators:

    $a ? $b : $c ? $d : $e

was being parsed as

    $a ? $b : ($c ? $d : $e)

but the original parser called for

    ($a ? $b : $c) ? $d : $e

Unfortunately this change breaks an oddity; we used to parse

    $a ? $b : $c = $d

as

    $a ? $b : ($c = $d)

but now parse it as

    ($a ? $b : $c) = $d

which obviously is not what was intended. I'll fix this in a later diff. (How, I'm not sure. The precedence rules of this language are messed up.)

Differential Revision: D4491210

fbshipit-source-id: 7b452a7538676e774b2ee03cba14b366162d5313