Allow named nested functions in FF parser
Summary:
PHP allows named nested function declarations in contexts where statements are expected. Hack does not in normal hh files, but does allow this in decl files.

In order to make the FF parser more compatible with php files, and to facilitate error reporting, the statement parser now detects non-anonymous function declarations where a statement is expected, and parses the function declaration.

In a later diff I'll make this an error in Hack files. (Though we might consider the feature; nested named functions are not a terrible feature.)

Differential Revision: D4491188

fbshipit-source-id: d35a8edd7d9b7a827f423e524067b56a2c2771bc